move bin packing code to a separate file
